clean up, add context menu to import/download templates
