+++ invokeai/frontend/web/src/services/api/endpoints/stylePresets.ts
@@ -92,13 +92,10 @@ export const stylePresetsApi = api.injectEndpoints({
       }),
       providesTags: ['FetchOnReconnect', { type: 'StylePreset', id: LIST_TAG }],
     }),
-    exportStylePresets: build.query<
-      string,
-      void
-    >({
+    exportStylePresets: build.query<string, void>({
       query: () => ({
-        url: buildStylePresetsUrl("/export"),
-        responseHandler: response => response.text()
+        url: buildStylePresetsUrl('/export'),
+        responseHandler: (response) => response.text(),
       }),
       providesTags: ['FetchOnReconnect', { type: 'StylePreset', id: LIST_TAG }],
     }),
